Created independent files for the home buttons and organized the directory
