+++ lib/features/home/components/home_buttons/hd_wallet_button.dart
@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../core/base_component.dart';
-import '../../../core/components/buttons.dart';
-import '../../../core/dialogs.dart';
-import '../../../core/providers/session_provider.dart';
-import '../../../utils/toast.dart';
-import '../../bridge/services/bridge_service.dart';
-import '../../encrypt/providers/wallet_is_encrypted_provider.dart';
-import '../../global_loader/global_loading_provider.dart';
+import '../../../../core/base_component.dart';
+import '../../../../core/components/buttons.dart';
+import '../../../../core/dialogs.dart';
+import '../../../../core/providers/session_provider.dart';
+import '../../../../utils/toast.dart';
+import '../../../bridge/services/bridge_service.dart';
+import '../../../encrypt/providers/wallet_is_encrypted_provider.dart';
+import '../../../global_loader/global_loading_provider.dart';
 
 class HdWalletButton extends BaseComponent {
   const HdWalletButton({
