+++ lib/features/home/screens/home_screen.dart
@@ -1,35 +1,24 @@
-import 'dart:io';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:process_run/shell.dart';
-import 'package:rbx_wallet/features/global_loader/global_loading_provider.dart';
+import 'package:rbx_wallet/features/home/components/home_buttons/mother_button.dart';
 import 'package:rbx_wallet/features/keygen/components/keygen_cta.dart'
     if (dart.library.io) 'package:rbx_wallet/features/keygen/components/keygen_cta_mock.dart';
-import 'package:rbx_wallet/features/mother/components/mother_button.dart';
-import 'package:rbx_wallet/features/smart_contracts/components/sc_creator/common/modal_container.dart';
-import 'package:rbx_wallet/features/validator/providers/validator_list_provider.dart';
 import 'package:rbx_wallet/features/wallet/components/wallet_selector.dart';
 import 'package:rbx_wallet/features/wallet/providers/wallet_list_provider.dart';
-import 'package:rbx_wallet/utils/files.dart';
-import 'package:rbx_wallet/utils/toast.dart';
 
 import '../../../core/base_screen.dart';
-import '../../../core/components/buttons.dart';
-import '../../../core/dialogs.dart';
-import '../../../core/env.dart';
 import '../../../core/providers/session_provider.dart';
-import '../../../core/theme/app_theme.dart';
-import '../../../core/utils.dart';
-import '../../bridge/models/log_entry.dart';
-import '../../bridge/providers/log_provider.dart';
-import '../../bridge/services/bridge_service.dart';
-import '../../encrypt/components/encrypt_wallet_button.dart';
-import '../../hd/components/hd_wallet_button.dart';
 import '../../hd/components/restore_hd_wallet_button.dart';
+import '../components/home_buttons/backup_button.dart';
+import '../components/home_buttons/encrypt_wallet_button.dart';
+import '../components/home_buttons/hd_wallet_button.dart';
+import '../components/home_buttons/open_db_button.dart';
+import '../components/home_buttons/open_log_button.dart';
+import '../components/home_buttons/print_addresses_button.dart';
+import '../components/home_buttons/print_validators_button.dart';
+import '../components/home_buttons/restart_cli_button.dart';
+import '../components/home_buttons/show_debug_data_button.dart';
 import '../components/log_window.dart';
 import '../components/transaction_window.dart';
 
@@ -60,7 +49,6 @@ class HomeScreen extends BaseScreen {
 
   @override
   Widget body(BuildContext context, WidgetRef ref) {
-    final cliStarted = ref.watch(sessionProvider).cliStarted;
     return SingleChildScrollView(
       child: Padding(
         padding: const EdgeInsets.all(8.0),
@@ -89,283 +77,17 @@ class HomeScreen extends BaseScreen {
                 spacing: 12.0,
                 runSpacing: 12.0,
                 children: [
-                  AppButton(
-                    label: "Print Addresses",
-                    icon: Icons.wallet,
-                    onPressed: !cliStarted
-                        ? null
-                        : () {
-                            final _log = ref.read(logProvider.notifier);
-
-                            _log.append(LogEntry(
-                              message: "Wallet Addresses:",
-                              variant: AppColorVariant.Secondary,
-                            ));
-
-                            final wallets = ref.read(walletListProvider);
-                            for (final wallet in wallets) {
-                              _log.append(LogEntry(
-                                message: "${wallet.address} (${wallet.balance} RBX)",
-                                variant: AppColorVariant.Success,
-                                textToCopy: wallet.address,
-                              ));
-                            }
-                          },
-                  ),
-                  AppButton(
-                    label: "Print Validators",
-                    icon: Icons.wifi,
-                    onPressed: !cliStarted
-                        ? null
-                        : () async {
-                            final _log = ref.read(logProvider.notifier);
-
-                            final validators = ref.read(validatorListProvider);
-
-                            if (validators.isEmpty) {
-                              _log.append(LogEntry(
-                                message: "No validators",
-                                variant: AppColorVariant.Danger,
-                              ));
-                              return;
-                            }
-                            _log.append(LogEntry(
-                              message: "Validators:",
-                              variant: AppColorVariant.Secondary,
-                            ));
-
-                            for (final validator in validators) {
-                              _log.append(
-                                LogEntry(
-                                  message: "${validator.address} => ${validator.isValidating ? 'Validating' : 'Not Validating'}",
-                                  variant: validator.isValidating ? AppColorVariant.Success : AppColorVariant.Info,
-                                  textToCopy: validator.address,
-                                ),
-                              );
-                            }
-                          },
-                  ),
-
+                  const PrintAdressesButton(),
+                  const PrintValidatorsButton(),
                   const MotherButton(),
                   const HdWalletButton(),
                   if (ref.watch(walletListProvider).isEmpty) const RestoreHdWalletButton(),
                   const EncryptWalletButton(),
-                  AppButton(
-                    label: "Show Debug Data",
-                    icon: Icons.analytics_outlined,
-                    onPressed: !cliStarted
-                        ? null
-                        : () async {
-                            final data = await BridgeService().getDebugInfo();
-                            InfoDialog.show(
-                              title: "Debug Data",
-                              content: Container(
-                                color: Colors.black54,
-                                child: Padding(
-                                  padding: const EdgeInsets.all(8.0),
-                                  child: SingleChildScrollView(
-                                    child: Column(
-                                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                                      mainAxisSize: MainAxisSize.min,
-                                      children: [
-                                        AppButton(
-                                          label: "Copy",
-                                          icon: Icons.copy,
-                                          variant: AppColorVariant.Success,
-                                          onPressed: () async {
-                                            await Clipboard.setData(ClipboardData(text: data));
-                                            Toast.message("Debug data copied to clipboard");
-                                          },
-                                        ),
-                                        const SizedBox(
-                                          height: 8,
-                                        ),
-                                        SelectableText(
-                                          data,
-                                          style: const TextStyle(
-                                            fontSize: 12,
-                                            fontFamily: "Courier",
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            );
-                          },
-                  ),
-                  if (!kIsWeb)
-                    AppButton(
-                      label: "Open DB Folder",
-                      icon: Icons.folder_open_rounded,
-                      onPressed: () async {
-                        // final shell = Shell(throwOnError: false);
-
-                        Directory appDocDir = await getApplicationDocumentsDirectory();
-                        String appDocPath = appDocDir.path;
-
-                        if (Platform.isMacOS) {
-                          appDocPath = appDocPath.replaceAll("/Documents", Env.isTestNet ? "/rbxtest" : "/rbx");
-                        } else {
-                          final winDir = await getApplicationSupportDirectory();
-                          appDocPath = winDir.path;
-                          appDocPath =
-                              appDocPath.replaceAll("\\Roaming\\com.example\\rbx_wallet_gui", "\\Local\\${Env.isTestNet ? 'RBXTest' : 'RBX'}");
-                        }
-
-                        openFile(File(appDocPath));
-
-                        // String cmd = "";
-                        // if (Platform.isMacOS) {
-                        //   appDocPath = appDocPath.replaceAll("/Documents", "/rbx");
-                        //   cmd = "open $appDocPath";
-                        // } else {
-                        //   appDocDir = await getApplicationSupportDirectory();
-
-                        //   appDocPath = appDocDir.path;
-
-                        //   appDocPath = appDocPath.replaceAll(
-                        //       "\\Roaming\\com.example\\rbx_wallet_gui",
-                        //       "\\Local\\rbx");
-                        //   cmd = "start $appDocPath";
-                        // }
-
-                        // shell.run(cmd);
-                      },
-                    ),
-                  if (!kIsWeb)
-                    AppButton(
-                      label: "Open Log",
-                      icon: Icons.open_in_new,
-                      onPressed: () async {
-                        final shell = Shell(throwOnError: false);
-
-                        Directory appDocDir = await getApplicationDocumentsDirectory();
-                        String appDocPath = appDocDir.path;
-
-                        String cmd = "";
-                        if (Platform.isMacOS) {
-                          appDocPath = appDocPath.replaceAll("/Documents", Env.isTestNet ? "/rbxtest" : "/rbx");
-                          cmd = "open $appDocPath/Databases${Env.isTestNet ? 'TestNet' : ''}/rbxlog.txt";
-                        } else {
-                          appDocDir = await getApplicationSupportDirectory();
-
-                          appDocPath = appDocDir.path;
-
-                          appDocPath = appDocPath.replaceAll("\\Roaming\\com.example\\rbx_wallet_gui", "\\Local\\RBX${Env.isTestNet ? 'Test' : ''}");
-                          cmd = "start $appDocPath\\Databases${Env.isTestNet ? 'TestNet' : ''}\\rbxlog.txt";
-                        }
-
-                        shell.run(cmd);
-                      },
-                    ),
-                  // if (Platform.isWindows)
-                  //   AppButton(
-                  //     label: "Clear Log",
-                  //     onPressed: () async {
-                  //       final confirmed = await ConfirmDialog.show(
-                  //         title: "Clear Log",
-                  //         body: "Are you sure you want to clear the log?",
-                  //         destructive: true,
-                  //         confirmText: "Clear",
-                  //         cancelText: "Cancel",
-                  //       );
-
-                  //       if (confirmed == true) {
-                  //         final success = await BridgeService().clearLog();
-                  //         if (success) {
-                  //           Toast.message("Log cleared");
-                  //         } else {
-                  //           Toast.error();
-                  //         }
-                  //       }
-                  //     },
-                  //   ),
-
-                  // if (Platform.isMacOS)
-                  AppButton(
-                    label: "Backup",
-                    icon: Icons.backup_outlined,
-                    onPressed: !cliStarted
-                        ? null
-                        : () async {
-                            showModalBottomSheet(
-                                backgroundColor: Colors.black87,
-                                // isScrollControlled: true,
-                                context: context,
-                                builder: (context) {
-                                  return ModalContainer(
-                                    color: Colors.black26,
-                                    withDecor: false,
-                                    children: [
-                                      ListTile(
-                                        title: const Text("Backup Keys"),
-                                        subtitle: const Text("Export and save your keys to a text file."),
-                                        leading: const Icon(Icons.wallet),
-                                        trailing: const Icon(Icons.chevron_right),
-                                        onTap: () async {
-                                          final success = await backupKeys(context, ref);
-                                          if (success == true) {
-                                            Navigator.of(context).pop();
-                                            Toast.message("Keys backed up successfully.");
-                                          } else {
-                                            Toast.error();
-                                          }
-                                        },
-                                      ),
-                                      ListTile(
-                                        title: const Text("Backup Media"),
-                                        subtitle: const Text("Zip and export your media assets."),
-                                        leading: const Icon(Icons.file_present),
-                                        trailing: const Icon(Icons.chevron_right),
-                                        onTap: () async {
-                                          final success = await backupMedia(context, ref);
-
-                                          if (success == true) {
-                                            Navigator.of(context).pop();
-                                            Toast.message("Media backed up successfully.");
-                                          } else {
-                                            Toast.error();
-                                          }
-                                        },
-                                      ),
-                                    ],
-                                  );
-                                });
-                          },
-                  ),
-
-                  AppButton(
-                    label: "Restart CLI",
-                    icon: Icons.restart_alt,
-                    // variant: AppColorVariant.Warning,
-                    onPressed: cliStarted
-                        ? () async {
-                            // BridgeService().killCli();
-
-                            final confirmed = await ConfirmDialog.show(
-                              title: "Restart",
-                              body: "Are you sure you want to restart the CLI?",
-                              confirmText: "Restart",
-                              cancelText: "Cancel",
-                              destructive: true,
-                            );
-
-                            if (confirmed) {
-                              ref.read(sessionProvider.notifier).restartCli();
-                            }
-                          }
-                        : null,
-                  ),
-
-                  // if (!kIsWeb && kDebugMode)
-                  //   AppButton(
-                  //     label: "Configure CLI",
-                  //     onPressed: () async {
-                  //       AutoRouter.of(context).push(const ConfigContainerScreenRoute());
-                  //     },
-                  //   ),
+                  const ShowDebugDataButton(),
+                  const OpenDbFolderButton(),
+                  const OpenLogButton(),
+                  const BackupButton(),
+                  const RestartCliButton(),
                 ],
               ),
             const Divider(),
